--- a/.github/workflows/wheel-builder.yml
+++ b/.github/workflows/wheel-builder.yml
@@ -66,7 +66,6 @@ jobs:
       matrix:
         PYTHON:
           - { VERSION: "cp311-cp311", ABI_VERSION: 'py37' }
-          - { VERSION: "cp311-cp311", ABI_VERSION: 'py39' }
           - { VERSION: "cp311-cp311", ABI_VERSION: 'py311' }
           - { VERSION: "pp310-pypy310_pp73" }
           - { VERSION: "pp311-pypy311_pp73" }
@@ -189,18 +188,6 @@ jobs:
             # build against
             _PYTHON_HOST_PLATFORM: 'macosx-10.9-universal2'
           - VERSION: '3.11'
-            ABI_VERSION: 'py39'
-            # Despite the name, this is built for the macOS 11 SDK on arm64 and 10.9+ on intel
-            DOWNLOAD_URL: 'https://www.python.org/ftp/python/3.11.3/python-3.11.3-macos11.pkg'
-            BIN_PATH: '/Library/Frameworks/Python.framework/Versions/3.11/bin/python3'
-            DEPLOYMENT_TARGET: '10.13'
-            # This archflags is default, but let's be explicit
-            ARCHFLAGS: '-arch x86_64 -arch arm64'
-            # See https://github.com/pypa/cibuildwheel/blob/c8876b5c54a6c6b08de5d4b1586906b56203bd9e/cibuildwheel/macos.py#L257-L269
-            # This will change in the future as we change the base Python we
-            # build against
-            _PYTHON_HOST_PLATFORM: 'macosx-10.9-universal2'
-          - VERSION: '3.11'
             ABI_VERSION: 'py311'
             # Despite the name, this is built for the macOS 11 SDK on arm64 and 10.9+ on intel
             DOWNLOAD_URL: 'https://www.python.org/ftp/python/3.11.3/python-3.11.3-macos11.pkg'
@@ -307,7 +294,6 @@ jobs:
           - {ARCH: 'x64', WINDOWS: 'win64', RUST_TRIPLE: 'x86_64-pc-windows-msvc'}
         PYTHON:
           - {VERSION: "3.11", "ABI_VERSION": "py37"}
-          - {VERSION: "3.11", "ABI_VERSION": "py39"}
           - {VERSION: "3.11", "ABI_VERSION": "py311"}
           - {VERSION: "pypy-3.10"}
           - {VERSION: "pypy-3.11"}