--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -61,6 +61,9 @@ Changelog
 * Added support for compiling against `aws-lc`_.
 * Parsing X.509 structures now more strictly enforces that ``Name`` structures
   do not have malformed ASN.1.
+* We now publish ``py311`` wheels that utilize the faster ``pyo3::buffer::PyBuffer``
+  interface, resulting in significantly improved performance for operations
+  involving small buffers.
 
 .. _v44-0-3: