--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,7 +21,8 @@ rust-version = "1.65.0"
 
 [workspace.dependencies]
 asn1 = { version = "0.21.1", default-features = false }
-pyo3 = { version = "0.24.2", features = ["abi3"] }
+pyo3 = { version = "0.24.2" }
+pyo3-build-config = { version = "0.24.2" }
 openssl = "0.10.72"
 openssl-sys = "0.9.108"