--- a/src/rust/build.rs
+++ b/src/rust/build.rs
@@ -6,9 +6,7 @@ use std::env;
 
 #[allow(clippy::unusual_byte_groupings)]
 fn main() {
-    for cfg in pyo3_build_config::get().build_script_outputs() {
-        println!("{cfg}");
-    }
+    pyo3_build_config::use_pyo3_cfgs();
 
     if let Ok(version) = env::var("DEP_OPENSSL_VERSION_NUMBER") {
         let version = u64::from_str_radix(&version, 16).unwrap();