--- a/src/rust/src/buf.rs
+++ b/src/rust/src/buf.rs
@@ -86,6 +86,9 @@ mod pybuffer_impl {
                 };
                 pyo3::exceptions::PyTypeError::new_err(errmsg)
             })?;
+            if bufobj.dimensions() != 1 {
+                return Err(pyo3::exceptions::PyTypeError::new_err("Only 1-dimensional byte buffers are supported."));
+            }            
             let len = bufobj.len_bytes();
             let buf = if len == 0 {
                 &[]