This pull request introduces support for optimized buffer handling using `pyo3::buffer::PyBuffer` on Python 3.11, along with related build system improvements.

- **Optimized buffer access**: On Python 3.11, the `CffiBuf` type now uses `pyo3::buffer::PyBuffer` to significantly speed up operations on small buffers.
- **New wheel builds**: Added publishing of `py311` wheels targeting `abi3-py311` using `pyo3::buffer`.

- **Removed the default `abi3` feature** from the `pyo3` dependency to allow per-build customization. This change ensures that users building locally will receive a wheel compatible with their target system, rather than being restricted to a `py37`-specific wheel. Users can still build a limited wheel with `pip wheel --config-settings="build-args=--features=pyo3/abi3-py37"`.
  **Rationale**: This adjustment ensures that downstream projects, like `piwheels` or Home Assistant, which build their own wheels, won't be unknowingly limited to an `abi3-py37` wheel and missing out on the performance improvements.
- **Added `pyo3-build-config`** as a build dependency to handle feature configuration cleanly.
- **Extended `noxfile.py`** to forward `--features=` args via `build-args` to `pip install`, enabling feature-specific builds in CI.
- **Updated `build.rs`** to print relevant PyO3 build script outputs.
- **Added `Py_3_11` cfg check** in Rust lint configuration to ensure feature-gated code is validated.

- Extended `.github/workflows/ci.yml` to test `rust,tests` jobs under different ABI3 feature flags to match the limited API used in the wheels.
  - `pyo3/abi3-py37`
  - `pyo3/abi3-py39`
  - `pyo3/abi3-py311`
- Extended `.github/workflows/wheel-builder.yml` to include `py311` ABI wheels for CPython 3.11.

- Documented the addition of `py311` wheels with improved buffer performance.

- Updated the documentation with instruction on how to build a wheel with the limited api.

| Benchmark | Base | PR | Delta |
| --------- | ---- | -- | ----- |
| `tests/bench/test_aead.py::test_aesccm_decrypt` | 2675.0 ns | 1480.25 ns | 45% faster |
| `tests/bench/test_aead.py::test_aesccm_encrypt` | 2665.0 ns | 1633.0 ns | 39% faster |
| `tests/bench/test_aead.py::test_aesgcm_decrypt` | 2895.0 ns | 1683.0 ns | 42% faster |
| `tests/bench/test_aead.py::test_aesgcm_encrypt` | 2785.0 ns | 1683.0 ns | 40% faster |
| `tests/bench/test_aead.py::test_aesocb3_decrypt` | 1867.0 ns | 1008.0 ns | 46% faster |
| `tests/bench/test_aead.py::test_aesocb3_encrypt` | 1964.0 ns | 1132.0 ns | 42% faster |
| `tests/bench/test_aead.py::test_aessiv_decrypt` | 1943.5 ns | 1440.25 ns | 26% faster |
| `tests/bench/test_aead.py::test_aessiv_encrypt` | 2034.0 ns | 1533.0 ns | 25% faster |
| `tests/bench/test_aead.py::test_chacha20poly1305_decrypt` | 3357.0 ns | 1893.0 ns | 44% faster |
| `tests/bench/test_aead.py::test_chacha20poly1305_encrypt` | 3346.0 ns | 1884.0 ns | 44% faster |
| `tests/bench/test_ec_load.py::test_load_ec_private_numbers` | 26379.0 ns | 26249.0 ns | 0% faster |
| `tests/bench/test_ec_load.py::test_load_ec_public_numbers` | 15629.0 ns | 15518.0 ns | 1% faster |
| `tests/bench/test_fernet.py::test_fernet_encrypt` | 20197.0 ns | 16811.0 ns | 17% faster |
| `tests/bench/test_hashes.py::test_sha256` | 2495.0 ns | 1994.0 ns | 20% faster |
| `tests/bench/test_hmac.py::test_hmac_sha256` | 3627.0 ns | 2605.0 ns | 28% faster |
| `tests/bench/test_x509.py::test_aki_public_bytes` | 631.0 ns | 641.0 ns | 2% slower |
| `tests/bench/test_x509.py::test_load_der_certificate` | 2985.0 ns | 2925.0 ns | 2% faster |
| `tests/bench/test_x509.py::test_load_pem_certificate` | 14557.0 ns | 14357.0 ns | 1% faster |
| `tests/bench/test_x509.py::test_object_identifier_constructor` | 325.1 ns | 314.6 ns | 3% faster |
| `tests/bench/test_x509.py::test_verify_docs_python_org` | 193217.5 ns | 183409.0 ns | 5% faster |Co-authored-by: pyca-boringbot[bot] <pyca-boringbot[bot]+106132319@users.noreply.github.com>…performance gains

This pull request introduces support for optimized buffer handling using `pyo3::buffer::PyBuffer` on Python 3.11, along with related build system improvements.

- **Optimized buffer access**: On Python 3.11, the `CffiBuf` type now uses `pyo3::buffer::PyBuffer` to significantly speed up operations on small buffers.
- **New wheel builds**: Added publishing of `py311` wheels targeting `abi3-py311` using `pyo3::buffer`.

- **Removed the default `abi3` feature** from the `pyo3` dependency to allow per-build customization. This change ensures that users building locally will receive a wheel compatible with their target system, rather than being restricted to a `py37`-specific wheel. Users can still build a limited wheel with `pip wheel --config-settings="build-args=--features=pyo3/abi3-py37"`.
  **Rationale**: This adjustment ensures that downstream projects, like `piwheels` or Home Assistant, which build their own wheels, won't be unknowingly limited to an `abi3-py37` wheel and missing out on the performance improvements.
- **Added `pyo3-build-config`** as a build dependency to handle feature configuration cleanly.
- **Extended `noxfile.py`** to forward `--features=` args via `build-args` to `pip install`, enabling feature-specific builds in CI.
- **Updated `build.rs`** to print relevant PyO3 build script outputs.
- **Added `Py_3_11` cfg check** in Rust lint configuration to ensure feature-gated code is validated.

- Extended `.github/workflows/ci.yml` to test `rust,tests` jobs under different ABI3 feature flags to match the limited API used in the wheels.
  - `pyo3/abi3-py37`
  - `pyo3/abi3-py39`
  - `pyo3/abi3-py311`
- Extended `.github/workflows/wheel-builder.yml` to include `py311` ABI wheels for CPython 3.11.

- Documented the addition of `py311` wheels with improved buffer performance.

- Updated the documentation with instruction on how to build a wheel with the limited api.

| Benchmark | Base | PR | Delta |
| --------- | ---- | -- | ----- |
| `tests/bench/test_aead.py::test_aesccm_decrypt` | 2675.0 ns | 1480.25 ns | 45% faster |
| `tests/bench/test_aead.py::test_aesccm_encrypt` | 2665.0 ns | 1633.0 ns | 39% faster |
| `tests/bench/test_aead.py::test_aesgcm_decrypt` | 2895.0 ns | 1683.0 ns | 42% faster |
| `tests/bench/test_aead.py::test_aesgcm_encrypt` | 2785.0 ns | 1683.0 ns | 40% faster |
| `tests/bench/test_aead.py::test_aesocb3_decrypt` | 1867.0 ns | 1008.0 ns | 46% faster |
| `tests/bench/test_aead.py::test_aesocb3_encrypt` | 1964.0 ns | 1132.0 ns | 42% faster |
| `tests/bench/test_aead.py::test_aessiv_decrypt` | 1943.5 ns | 1440.25 ns | 26% faster |
| `tests/bench/test_aead.py::test_aessiv_encrypt` | 2034.0 ns | 1533.0 ns | 25% faster |
| `tests/bench/test_aead.py::test_chacha20poly1305_decrypt` | 3357.0 ns | 1893.0 ns | 44% faster |
| `tests/bench/test_aead.py::test_chacha20poly1305_encrypt` | 3346.0 ns | 1884.0 ns | 44% faster |
| `tests/bench/test_ec_load.py::test_load_ec_private_numbers` | 26379.0 ns | 26249.0 ns | 0% faster |
| `tests/bench/test_ec_load.py::test_load_ec_public_numbers` | 15629.0 ns | 15518.0 ns | 1% faster |
| `tests/bench/test_fernet.py::test_fernet_encrypt` | 20197.0 ns | 16811.0 ns | 17% faster |
| `tests/bench/test_hashes.py::test_sha256` | 2495.0 ns | 1994.0 ns | 20% faster |
| `tests/bench/test_hmac.py::test_hmac_sha256` | 3627.0 ns | 2605.0 ns | 28% faster |
| `tests/bench/test_x509.py::test_aki_public_bytes` | 631.0 ns | 641.0 ns | 2% slower |
| `tests/bench/test_x509.py::test_load_der_certificate` | 2985.0 ns | 2925.0 ns | 2% faster |
| `tests/bench/test_x509.py::test_load_pem_certificate` | 14557.0 ns | 14357.0 ns | 1% faster |
| `tests/bench/test_x509.py::test_object_identifier_constructor` | 325.1 ns | 314.6 ns | 3% faster |
| `tests/bench/test_x509.py::test_verify_docs_python_org` | 193217.5 ns | 183409.0 ns | 5% faster |